fix: inline local PDF helpers to avoid missing-module crash in packaged app
